# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Messenger backend api",
        "description": "Messenger backend",
        "contact": {
            "email": "tcom.developer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://crypt-talk.test/api",
            "description": "Dev server"
        },
        {
            "url": "http://95.217.156.195/api",
            "description": "Dev server"
        }
    ],
    "paths": {
...
        "/user/search": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Search user by name or email",
                "operationId": "7495095e5b6c23625c419daaa0200f1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "search": {
                                        "description": "search string",
                                        "type": "string",
                                        "default": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/show/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Find user by ID",
                "operationId": "2359a38d992595bc4001f41c0b92d89d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/edit": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Find user by auth token",
                "operationId": "fef4feace1e1197ecf512a20eb5bf906",
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update user data",
                "operationId": "25fa6e59311f61c2be4cca80c602dde3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "about",
                                    "avatar"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "about": {
                                        "description": "About user",
                                        "type": "string",
                                        "example": "Software developer from Belgrade"
                                    },
                                    "avatar": {
                                        "description": "User avatar file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/change-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change user password",
                "operationId": "48fbc004e5b7e15198e50b6409fbd7ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "about",
                                    "avatar"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "Password123#"
                                    },
                                    "password2": {
                                        "description": "User password repeated",
                                        "type": "string",
                                        "example": "Password123#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorize with bearer token",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
}